Install MT settings override in tcms_settings_dir/
this will make it very easy to configure MT instances by
providing all of the static configuration in the package and
installing it where Kiwi TCMS can find it.

ATM one only needs to specify KIWI_TENANTS_DOMAIN env variable!

Use the new settings file for testing and debugging so we can
have configuration only in a single place.