Always lower case schema_name
because Postgres and django_tenants allow upper case letters as
schema_name but that will not play well as a host name. Same
thing goes for `_` where it is allowed in DNS records and local
domain name but not as part of a hostname (e.g. FQDN). See
https://github.com/tomturner/django-tenants/issues/385

Ideally this will be resolved upstream!