Replace ModelChoiceField with UserField. Fixes #114
this is a custom field which accepts username, email or ID and
is used in many places in Kiwi TCMS. Here it is used to avoid
presenting a huge list of existing accounts to the administrator
and making the selection a bit less clunky.

Still doesn't implement an auto-complete widget but that's
unrelated.