More robust tenant domain detection
This is used on the login page to construct the URLs for social login
buttons. Usually important only for 'public' schema.

In the case where public tenant's domain doesn't use a prefix (e.g. matches
KIWI_TENANT_DOMAIN) the previous behavior was leading to incorrect
results by always adding the "public." prefix!

With this change the actual domain from the DB will be returned.

When creating new tenants this function is used to automaticaly assign a
domain name to them, which is always going to use the prefix format b/c
there's no existing Tenant or Domain in the DB.